Damage numbers in battle
Adds an option to display the amount of damage done on hit in battle, Live A Live/Pokerogue style.
This number does not adjust itself based on max HP or anything like that; it just displays the value of gBattleMoveDamage. This means it will give the player info about the target's HP that they don't normally have.
The constants file includes config options for the feature: you can change how long the number stays on screen for, the delay between each number showing up (including removing the delay), make the numbers only appear from direct attacks, and choose whether or not the numbers change color based on the type effectiveness of the attack.

I haven't thoroughly tested this for edge cases, so there may be some unexpected behavior if you call these battle script commands in an unexpected way. e.g. I don't know if it will play nice with expansion. But it's compact, so errors should be fixable.